devide backend logic into routes and controllers
