+++ backend/controllers/UserController.js
@@ -1,0 +1,130 @@
+const express = require('express');
+const router = express.Router();
+const gravatar = require('gravatar');
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const passport = require('passport');
+const validateRegisterInput = require('../validation/register');
+const validateLoginInput = require('../validation/login');
+
+const Users = require('../models/User');
+
+exports.registerUser = function (req, res) {
+    const {errors, isValid} = validateRegisterInput(req.body);
+
+    if (!isValid) {
+        return res.status(400).json(errors);
+    }
+    Users.findOne({
+        email: req.body.email
+    }).then(user => {
+        if (user) {
+            return res.status(400).json({
+                email: 'Email already exists'
+            });
+        } else {
+            const newUser = new Users({
+                name: req.body.name,
+                email: req.body.email,
+                password: req.body.password,
+                isAdmin: req.body.isAdmin
+            });
+
+            bcrypt.genSalt(10, (err, salt) => {
+                if (err) console.error('There was an error', err);
+                else {
+                    bcrypt.hash(newUser.password, salt, (err, hash) => {
+                        if (err) console.error('There was an error', err);
+                        else {
+                            newUser.password = hash;
+                            newUser
+                                .save()
+                                .then(user => {
+                                    res.json(user)
+                                });
+                        }
+                    });
+                }
+            });
+        }
+    });
+}
+
+exports.loginUser = function (req, res) {
+    const {errors, isValid} = validateLoginInput(req.body);
+
+    if (!isValid) {
+        return res.status(400).json(errors);
+    }
+
+    const email = req.body.email;
+    const password = req.body.password;
+
+    Users.findOne({email})
+        .then(user => {
+            if (!user) {
+                errors.email = 'User not found'
+                return res.status(404).json(errors);
+            }
+            bcrypt.compare(password, user.password)
+                .then(isMatch => {
+                    if (isMatch) {
+                        const payload = {
+                            id: user.id,
+                            name: user.name,
+                            // avatar: user.avatar,
+                            isAdmin: user.isAdmin
+                        }
+                        jwt.sign(payload, 'secret', {
+                            expiresIn: 3600
+                        }, (err, token) => {
+                            if (err) console.error('There is some error in token', err);
+                            else {
+                                res.json({
+                                    success: true,
+                                    token: `Bearer ${token}`
+                                });
+                            }
+                        });
+                    } else {
+                        errors.password = 'Incorrect Password';
+                        return res.status(400).json(errors);
+                    }
+                });
+        });
+}
+
+exports.editUser = function (req, res, next) {
+    Users.findById(req.params.id, (error, data) => {
+        if (error) {
+            return next(error)
+        } else {
+            res.json(data)
+        }
+    })
+}
+
+exports.updateUser = function (req, res, next) {
+    Users.findByIdAndUpdate(req.params.id, {
+        $set: req.body
+    }, (error, data) => {
+        if (error) {
+            return next(error);
+            console.log(error);
+        } else {
+            res.json(data)
+            console.log('Profile updated successfully !')
+        }
+    })
+}
+
+exports.authenticateMe = function (req, res) {
+    return res.json({
+        id: req.user.id,
+        name: req.user.name,
+        email: req.user.email,
+        isAdmin: req.user.isAdmin
+    });
+}
+
+
