+++ backend/routes/user-routes.js
@@ -1,140 +1,19 @@
 const express = require('express');
 const router = express.Router();
-const gravatar = require('gravatar');
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
+var user_Controller = require('../controllers/UserController');
 const passport = require('passport');
-const validateRegisterInput = require('../validation/register');
-const validateLoginInput = require('../validation/login');
 
-const User = require('../models/User');
 
-router.post('/register', function (req, res) {
-    const {errors, isValid} = validateRegisterInput(req.body);
+router.route('/register').post(user_Controller.registerUser);
 
-    if (!isValid) {
-        return res.status(400).json(errors);
-    }
-    User.findOne({
-        email: req.body.email
-    }).then(user => {
-        if (user) {
-            return res.status(400).json({
-                email: 'Email already exists'
-            });
-        } else {
-            // const avatar = gravatar.url(req.body.email, {
-            //     s: '200',
-            //     r: 'pg',
-            //     d: 'mm'
-            // });
-            const newUser = new User({
-                name: req.body.name,
-                email: req.body.email,
-                password: req.body.password,
-                // avatar,
-                isAdmin: req.body.isAdmin
-            });
-
-            bcrypt.genSalt(10, (err, salt) => {
-                if (err) console.error('There was an error', err);
-                else {
-                    bcrypt.hash(newUser.password, salt, (err, hash) => {
-                        if (err) console.error('There was an error', err);
-                        else {
-                            newUser.password = hash;
-                            newUser
-                                .save()
-                                .then(user => {
-                                    res.json(user)
-                                });
-                        }
-                    });
-                }
-            });
-        }
-    });
-});
-
-router.post('/login', (req, res) => {
-
-    const {errors, isValid} = validateLoginInput(req.body);
-
-    if (!isValid) {
-        return res.status(400).json(errors);
-    }
-
-    const email = req.body.email;
-    const password = req.body.password;
-
-    User.findOne({email})
-        .then(user => {
-            if (!user) {
-                errors.email = 'User not found'
-                return res.status(404).json(errors);
-            }
-            bcrypt.compare(password, user.password)
-                .then(isMatch => {
-                    if (isMatch) {
-                        const payload = {
-                            id: user.id,
-                            name: user.name,
-                            // avatar: user.avatar,
-                            isAdmin: user.isAdmin
-                        }
-                        jwt.sign(payload, 'secret', {
-                            expiresIn: 3600
-                        }, (err, token) => {
-                            if (err) console.error('There is some error in token', err);
-                            else {
-                                res.json({
-                                    success: true,
-                                    token: `Bearer ${token}`
-                                });
-                            }
-                        });
-                    } else {
-                        errors.password = 'Incorrect Password';
-                        return res.status(400).json(errors);
-                    }
-                });
-        });
-});
-
+router.route('/login').post(user_Controller.loginUser);
 
 //Get profile
-router.route('/edit-profile/:id').get((req, res, next) => {
-    User.findById(req.params.id, (error, data) => {
-        if (error) {
-            return next(error)
-        } else {
-            res.json(data)
-        }
-    })
-})
+router.route('/edit-profile/:id').get(user_Controller.editUser);
 
 // Update Profile
-router.route('/update-profile/:id').put((req, res, next) => {
-    User.findByIdAndUpdate(req.params.id, {
-        $set: req.body
-    }, (error, data) => {
-        if (error) {
-            return next(error);
-            console.log(error);
-        } else {
-            res.json(data)
-            console.log('Profile updated successfully !')
-        }
-    })
-})
+router.route('/update-profile/:id').put(user_Controller.updateUser);
 
-router.get('/me', passport.authenticate('jwt', {session: false}), (req, res) => {
-    return res.json({
-        id: req.user.id,
-        name: req.user.name,
-        email: req.user.email,
-        isAdmin: req.user.isAdmin
-    });
-});
+router.get('/me', passport.authenticate('jwt', {session: false}), user_Controller.authenticateMe);
 
 module.exports = router;
